Minified & tuto
New assets ! The tuorial looks finished for now, with the beautiful
arrows.
I added a sound for the gain of dollars, but I'm not sure it's good.
Using of minified versions of jQuery and Crafty, much faster to load the
game now.
The SCORE mode permits us to not update the score at the beginning when
we add X lines.